Remove extraneous compiler warnings: * Turn on /LTCG on release builds (the release builds may take longer) * ignore the LNK4099 error "pdb" missing on the third-party lib files (which were all compiled in Release configuration so there's no debug info) * updated some code for explicit casting so compiler doesn't complain

+++ src/JPEGView/AVIFWrapper.cpp
@@ -76,7 +76,7 @@ void* AvifReader::ReadImage(int& width,
 	height = cache.rgb.height;
 	has_animation = cache.decoder->imageCount > 1;
 	frame_count = cache.decoder->imageCount;
-	frame_time = cache.decoder->imageTiming.duration * 1000;
+	frame_time = (int)(cache.decoder->imageTiming.duration * 1000.0);
 
 	uint8_t* pixels = cache.rgb.pixels;
 	size_t size = width * nchannels * height;
